fix: mydashboard api 연동 수정
